Filestructure changed (includes code refactoring)
