+++ data/sqlite/sqlite3_analyzer.exe
