+++ data/xml/inventoryEntries.xsd
