+++ src/controller/Main.java
@@ -1,4 +1,4 @@
-package logic;
+package controller;
 
 public class Main {
