--- src/database/inhouse/Database.java
@@ -1,9 +1,0 @@
-package database.inhouse;
-
-
-import database.DatabaseInterface;
-
-public class Database implements DatabaseInterface {
-
-
-}
