+++ src/model/Category.java
@@ -1,4 +1,4 @@
-package models;
+package model;
 
 public class Category {
 	private int id;
