+++ src/model/Product.java
@@ -1,4 +1,4 @@
-package models;
+package model;
 
 public class Product {
 	private int id;
